Update renderers to tp4 of web3.js (#110)

+++ packages/renderers-rust/src/utils/codecs.ts
@@ -4,13 +4,13 @@ import { getBase16Encoder, getBase58Encoder, getBase64Encoder, getUtf8Encoder } 
 export function getBytesFromBytesValueNode(node: BytesValueNode): Uint8Array {
     switch (node.encoding) {
         case 'utf8':
-            return getUtf8Encoder().encode(node.data);
+            return getUtf8Encoder().encode(node.data) as Uint8Array;
         case 'base16':
-            return getBase16Encoder().encode(node.data);
+            return getBase16Encoder().encode(node.data) as Uint8Array;
         case 'base58':
-            return getBase58Encoder().encode(node.data);
+            return getBase58Encoder().encode(node.data) as Uint8Array;
         case 'base64':
         default:
-            return getBase64Encoder().encode(node.data);
+            return getBase64Encoder().encode(node.data) as Uint8Array;
     }
 }
